refactor: extract usePublicRoutes to a store-free module
The appbar reads usePublicRoutes on every route. Sourcing it from
applicationHooks dragged that module's slice/API imports into every client
bundle, including the public routes. Move the hook to a module free of Redux
and feature imports so the shell no longer pulls the DJ feature graph.
